requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==4.1.0
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.5
cssselect==1.1.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
fake-useragent==0.1.11
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.17.2
Jinja2==3.0.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.1
lxml==4.9.1
MarkupSafe==2.0.1
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.19.5
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parse==1.19.0
parso==0.7.1
pickleshare==0.7.5
prometheus-client==0.14.1
prompt-toolkit==3.0.30
pycparser==2.21
pyee==8.2.2
Pygments==2.12.0
pyparsing==3.0.9
pyppeteer==0.2.6
pyquery==1.4.3
pyrsistent==0.18.0
python-dateutil==2.8.2
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==1.1.6
pyzmq==23.2.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
requests-html==0.10.0
selenium==3.141.0
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
terminado==0.12.1
testpath==0.6.0
tornado==6.1
tqdm==4.64.0
traitlets==4.3.3
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==9.1
widgetsnbextension==3.6.1
wincertstore==0.2
zipp==3.6.0
